pyproject.toml
[project]
name = "pymdrive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jgobbo", email = "59982706+jgobbo@users.noreply.github.com" }
]
requires-python = ">=3.9"
dependencies = [
    "pyserial>=3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
